refactor: move plugin to separate file - export all important types
